+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Feature to folder can be mapped with file extension .feature-folder and .feature-to-folder
 - Feature to file can be mapped with file extension .feature-file and .feature-to-file
 
+### Added
+- Added a View of the metrics to the project
+
 ## [0.0.4] - 2024-03-06
 
 ### Added